backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
SQLAlchemy==2.0.35
psycopg[binary]==3.2.10
python-dotenv==1.0.1
